Added SccGroups tests and organized tests
