+++ src/algo/mod.rs
@@ -1,87 +1,20 @@
-use std::rc::Rc;
-
 use crate::graph_impl::DiGraph;
+use crate::js_helpers::SccGroups;
 use crate::GraphError;
 use petgraph::algo;
 use petgraph::graph;
 use wasm_bindgen::prelude::*;
 
-#[wasm_bindgen]
-pub struct SccGroups {
-    #[wasm_bindgen(skip)]
-    pub inner: Rc<Vec<Vec<graph::NodeIndex>>>,
-}
-
-#[wasm_bindgen]
-pub struct SccGroup {
-    #[wasm_bindgen(skip)]
-    pub inner: Rc<Vec<Vec<graph::NodeIndex>>>,
-    #[wasm_bindgen(readonly)]
-    pub index: usize
-}
-
-#[wasm_bindgen]
-impl SccGroups {
-    fn new(val: Vec<Vec<graph::NodeIndex>>) -> Self {
-        Self { inner: Rc::new(val) }
-    }
-
-    pub fn get(&self, index: usize) -> JsValue {
-        match self.inner.len() > index {
-            true => JsValue::from(SccGroup::new(self.inner.clone(), index)),
-            false => JsValue::undefined(),
-        }
-    }
-}
-
-impl SccGroups {
-    pub fn to_std_vec(&self) -> Vec<Vec<usize>> {
-        self.inner.iter().map(|group| group.iter().map(|item| item.index()).collect()).collect()
-    }
-}
-
-
-#[wasm_bindgen]
-impl SccGroup {
-    fn new(inner: Rc<Vec<Vec<graph::NodeIndex>>>, index: usize) -> Self {
-        Self { inner: inner, index }
-    }
-
-    pub fn get(&self, index: usize) -> JsValue {
-        match self.inner.get(self.index) {
-            None => panic!("INTERNAL ERROR IN SCCGROUP"),
-            Some(scc_group) => match scc_group.get(index) {
-                Some(node_index) => JsValue::from(node_index.index() as u32),
-                None => JsValue::undefined(),
-            }
-        }
-    }
-
-    pub fn get_all(&self, index: usize) -> JsValue {
-        match self.inner.get(self.index) {
-            None => panic!("INTERNAL ERROR IN SCCGROUP"),
-            Some(scc_group) => match scc_group.get(index) {
-                Some(node_index) => JsValue::from(node_index.index() as u32),
-                None => JsValue::undefined(),
-            }
-        }
-    }
-}
-
-//fn helper2(nodeIndexVec: Vec<graph::NodeIndex>) -> Vec<u32> {
-//    nodeIndexVec.iter().map(|item| item.index() as u32).collect()
-//}
-
 #[wasm_bindgen(js_name = kosarajuScc)]
 pub fn kosaraju_scc(graph: &DiGraph) -> SccGroups {
     let scc_groups: Vec<Vec<graph::NodeIndex>> = algo::kosaraju_scc(&graph.graph);
-    return SccGroups::new(scc_groups)
+    return SccGroups::new(scc_groups);
 }
 
-#[wasm_bindgen(js_name = tarjan_scc)]
+#[wasm_bindgen(js_name = tarjanScc)]
 pub fn tarjan_scc(graph: &DiGraph) -> SccGroups {
     let scc_groups: Vec<Vec<graph::NodeIndex>> = algo::tarjan_scc(&graph.graph);
-    return SccGroups::new(scc_groups)
+    return SccGroups::new(scc_groups);
 }
 
 #[wasm_bindgen]
@@ -98,5 +31,45 @@ pub fn toposort(graph: &DiGraph) -> Result<Vec<usize>, JsValue> {
             Ok(graph_error_js) => Result::Err(graph_error_js),
             Err(failure_msg) => Result::Err(JsValue::from_str(&failure_msg)),
         },
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::js_helpers::test::new_test_graph;
+    use crate::{GraphError, GraphItemType};
+    use wasm_bindgen_test::*;
+
+    #[wasm_bindgen_test]
+    fn can_sort_nodes() {
+        let (g, _nodes, _edges) = new_test_graph();
+        let sorted_nodes = toposort(&g).unwrap();
+        assert_eq!(sorted_nodes, vec![4, 1, 0, 3, 2]);
+    }
+
+    #[wasm_bindgen_test]
+    fn can_detect_cycles() {
+        let (mut g, _nodes, _edges) = new_test_graph();
+        g.add_edge(2, 1, JsValue::NULL);
+        let sort_err = toposort(&g).unwrap_err();
+        let expect_err = GraphError::new("Cycle detected", GraphItemType::Node, 1);
+        assert_eq!(sort_err.into_serde::<GraphError>().unwrap(), expect_err);
+    }
+
+    #[wasm_bindgen_test]
+    fn can_get_kosaraju_scc() {
+        let (mut g, _nodes, _edges) = new_test_graph();
+        g.add_edge(2, 1, JsValue::NULL);
+        let scc_groups: SccGroups = kosaraju_scc(&g);
+        assert_eq!(scc_groups.copy_to_std_vec(), vec![vec![0, 3, 2, 1], vec![4]]);
+    }
+
+    #[wasm_bindgen_test]
+    fn can_get_tarjan_scc() {
+        let (mut g, _nodes, _edges) = new_test_graph();
+        g.add_edge(2, 1, JsValue::NULL);
+        let scc_groups: SccGroups = tarjan_scc(&g);
+        assert_eq!(scc_groups.copy_to_std_vec(), vec![vec![3, 1, 2, 0], vec![4]]);
     }
 }
